requirements.txt
asgiref==3.4.1
astroid==2.6.6
backcall==0.2.0
decorator==5.0.9
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
ipython==7.26.0
ipython-genutils==0.2.0
isort==5.9.3
jedi==0.18.0
lazy-object-proxy==1.6.0
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.10.0
pylint==2.9.6
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
wrapt==1.12.1
